requirements.txt
sanic
sanic-cors
gunicorn
tensorflow
git+git://github.com/minimaxir/textgenrnn.git@master#egg=textgenrnn
google-cloud-storage